requirements.txt
slack-sdk>=3.21.0
python-dotenv>=1.0.0
regex>=2023.10.3
python-dateutil>=2.8.2
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.1 
python-dotenv>=1.0.0 
matplotlib>=3.5.0
slack-bolt>=1.14.0 
